backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
tortoise-orm[asyncmy]==0.20.0
aiomysql==0.2.0    # 生产环境

# 认证和安全
passlib[bcrypt]==1.7.4

# Redis和缓存
aioredis==2.0.1

# 异步任务
celery==5.3.4

# HTTP客户端
httpx==0.25.2

# 日志
loguru==0.7.2

# 数据验证
pydantic==2.5.0

# 配置管理
pydantic-settings==2.1.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
python-multipart==0.0.6
python-jose[cryptography]==3.3.0